+++ NEWS.md
@@ -9,6 +9,13 @@ editor_options:
 Add support for StellarNet, Brewer, and Bentham spectrometers (examples
 of data files are welcome for these and other instruments).
 
+## photobiologyInOut 0.4.33
+
+- Add function `read_asd_tsv()` for importing spectral energy irradiance,
+reflectance or transmittance data acquired with ASD spectrometers and converted
+from binary to text with software from ASD.
+- Update the vignette.
+
 ## photobiologyInOut 0.4.32
 
 - Add function `read_oo_ovirrad()` for importing spectral energy irradiance data